server/Inventory-service/requirements.txt
# Flask==2.3.3
# Flask-RESTful==0.3.9
# psycopg2-binary==2.9.11
# pytest==7.4.0
# flask-jwt-extended

# server/inventory-service/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
